structuring the webstie
added a two pictures of boston.
not required but my not, one is a nice picture of boston, other is just me sitting in traffic on i-90 W.
made some rearrangements to the website code to properly show the images.
replaced combo chart with two separate charts to reduce visual noise.